moving my multi-sample analysis code over
